+++ packages/admin/src/components/PublishDraftModal/index.jsx
@@ -1,6 +1,7 @@
 import { publishDraft } from '@/services/van-blog/api';
 import { Modal, ModalForm, ProFormSelect, ProFormText } from '@ant-design/pro-components';
 import { message } from 'antd';
+import PathnameField from '../PathnameField';
 export default function (props) {
   const { title, id, trigger, action, onFinish } = props;
   return (
@@ -67,14 +68,7 @@ export default function (props) {
             autocomplete: 'new-password',
           }}
         />
-        <ProFormText
-          width="md"
-          id="pathname"
-          name="pathname"
-          label="自定义路径名"
-          tooltip="文章发布后的路径将为 /post/[自定义路径名]，如果未设置则使用文章 id 作为路径名"
-          placeholder="留空或为空则使用 id 作为路径名"
-        />
+        <PathnameField />
         <ProFormText.Password
           label="密码"
           width="md"